requirements.txt
alembic==1.9.2
aniso8601==9.0.1
apispec==6.0.2
apispec-webframeworks==0.5.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Ext==0.1
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==4.0.2
Flask-RESTful==0.3.9
flask-restful-swagger==0.20.2
Flask-SQLAlchemy==3.0.2
flask-swagger-ui==4.11.1
greenlet==2.0.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
